backend/requirements.txt
fastapi[standard]
uvicorn
psycopg2-binary
sqlalchemy
pydantic-settings
tqdm
asyncpg